@@ -20,7 +20,7 @@
   "devDependencies": {
     "@types/node": "^11.9.5",
     "tslint": "^5.19.0",
-    "typescript": "^3.6.2"
+    "typescript": "^3.9.7"
   },
   "dependencies": {
     "sqrl": "^0.6.4"
