@@ -43,7 +43,7 @@
     "sqrl-load-functions": "^0.6.4",
     "ts-jest": "^24.0.2",
     "ts-node": "^8.3.0",
-    "typescript": "^3.6.2"
+    "typescript": "^3.9.7"
   },
   "engines": {
     "node": ">=6.0.0"
